requirements.txt
requests==2.31.0
python-dotenv==1.0.0
rich==13.7.0
opencv-python==4.8.1.78
numpy==1.24.3
Pillow==10.2.0
ultralytics==8.1.2
torch==2.2.0
transformers==4.37.2 